Additional rework of subranges and their memory sizes.
Modula2 subranges specify a sized base type.  They need to be
specified in size when used in unsafe writing and reading of array
elements in tables, but subranges when used as array bounds and for
general range checking.

This fix uses BITS to set sizes when used as elements.  This approach
failed earlier, because it created multiple levels, i.e.: BITS FOR
BITS FOR, which crashed the M3 compiler.  This seems to have now gone
away.