Merge common fields of flattened variant record Tree0.yyNode.
Originally a Modula2 variant record, several fields were
duplicated at the same displacement in different variants,
thus being overlaid in Modula2.  m2tom3's flattening put
each variant in a different inner record, eliminating
the overlaying.  A NIL field value in Modula3 version
was causing a segfault, when no change was expected.  So
this was a hopeful strategy to reproduce the overlaying.
It stopped this segfault, so probably was correct.