Unravel instantiations of ExpArrays.
It seems impossible to avoid unsafe propagating everywhere,
through interfaces, once there is one unsafe iterface.  If
you want to expand an array with anything in it declared in
an unsafe interface, you are doomed.

Could Automaton.i3 be divided such that everything involved
in an expandable array is in a separate, safe interface?