Rename types in Parser, for less confusion.
These types are converted from Modula-2 and are quite fragile.
On the one hand, some of them, in some contexts, need to be
subranges (particularly, as fixed array subscript types),
and other times, need to have the same size as in Modula-2, to avoid
undermining various unsafe address arithmetic.  Modula-3 infers
its own sizes from subranges, except for fields and elements when
BITS FOR is used.  But assignments involving scalars with BITS FOR
types present problems and even CM3 code generator failures.

So, if it is a BITS FOR type, its name ends in "Packed", otherwise
not, the relevant ones ending in "Range".

Additionally, CM3 has code generator failures assigning between two
BITS FOR types, at times.  Actual cases where this has happened are
replaced by two-step copies with an intermediate, unpacked temporary.