Rework error handling.
This was a mess and still is to a great extent, mostly because
of extensive use of highly unsafe coding.  Error messages have
a parameter with dynamically-multiple alternative types, all
passed as ADDRESS, with a separate code to denote what it really
is.  Most or all of these are not inherently pointers, those are
added by ADDRESS when passing.  Giving them atype and dereferencing
involves using BYTESIZE on the referent.  Some of these can change
size, with little hint to the changer that it will undermine this
mechanism.  It looks like lalr, at least, hasn't make any such
size changes, but some, like state, could be candidates.
Caviat emptor.

The motive for the change is use of IntSets instead of Cocktail
Set.  The latter is a traced ref, while the former is untraced.
Trying to cram these into a single type would be a nightmare.
So there is some duplication of higher-level procedures and
everything lower has two alternative params/fields, one ADDRESS,
the other REFANY.  Only one of these is used for a given message.

This also required gettin IntSets into front, with all its
entourage.  That in turn has to be removed from lalr, which will
be a separate commit.