Some rework of FrontErrors.
Error classes and codes were not being interpreted.
Replaced the read-in table by coded-in CASE stmts.

Also added flushes for immediate messages.

This whole thing needs to be reworked for FM3-style messages,
but that is messy, because changes need to be made in the
FM3 parser, that may not be good for lalr in general. And
the parser skeleton is part of general lalr.