Fix nonterminal bias bug in generated parser.
How this ever worked in original lalr baffles me. Or did it?
Nonterminal increments to yyNComb elements were using absolute
NT values, but yyNComb is generated to use values based from the
lowest-numbered nonterminal.  This fix makes parsing get farther.

NOTE: Unrelatedly, yyNComb itself also has lower bound of lowest
nonterminal, but its subscripts are not states, symbols, or anything
else, just places pointed to by [trueState, nonterm] pairs.  It
would be much cleaner to use zero for its lower bound.  In the
parser itself, subscripts to yyNComb are not used anyway (except
as comments in the array constructor, which I just added), in
favor of unsafe pointers directly to elements, found in yyNBase.

Also a number of renames and mostly just comments, especially
to reduce the confusion caused by calling parsing steps "States",
when some of these are states, but others are actions, denoted
as biased production numbers, above the space of true states.
A true state corresponds to an item set.  These higher numbered
actions do not.  This is clearer if you first read "Lalr - A
Generator for Efficient Parsers".