Fix subtle bug in conversion of Select.
Sets.Select returns 0 when set is empty.  This is ambiguous
(could mean 0 is a member) and incorrect, but other code depends
on this behavour, specifically Continue.MakeContinuation.

Intsets, used instead of Cocktail Sets, in the M3 conversion returns
a reserved value of maximally negative integer, which choked MakeContinuation.

Added Sets.SelectIntSets and ExtractIntSets that change the IntSets
behavour to the orignal Sets behavour.  Use SelectIntSets in MakeContinuation.